DAT COMMIT THO
On a du temps , des feux, et des chars qui run partout lol

+++ SPTR/SPTR/Cellule.cs
@@ -38,12 +38,12 @@ namespace SPTR
 
         virtual public int CoordonneeXEchelle
         {
-            get { return CoordonneeX * TailleCellule; }
+            get { return (int)CoordonneeX * TailleCellule; }
         }
 
         virtual public int CoordonneeYEchelle
         {
-            get { return CoordonneeY * TailleCellule; }
+            get { return (int)CoordonneeY * TailleCellule; }
         }
 
     }
